backend/requirements.txt
requests
python-dotenv
flask
groq
flask-cors
